+++ desktop/applis/apps/gui/gui_games/cardgamesgui/src/main/java/cards/gui/dialogs/EditorTarot.java
@@ -360,7 +360,7 @@ public final class EditorTarot extends DialogTarot implements SetterSelectedCard
             nombreCartesSelectionnees=0;
         } else {
             String mes_ = getMessages().getVal(ERROR_MOVE);
-            mes_ = StringUtil.simpleStringsFormat(mes_, Long.toString(m.total()), Long.toString(max_-taille_), listeTwo.getSelectedComboItem());
+            mes_ = StringUtil.simpleStringsFormat(mes_, Long.toString(m.total()), Long.toString((long)max_-taille_), listeTwo.getSelectedComboItem());
             ConfirmDialog.showMessage(this, mes_, getMessages().getVal(ERROR_MOVE_TITLE), lg_, JOptionPane.ERROR_MESSAGE, getMain().getConfirmDialog());
             //JOptionPane.showMessageDialog(this,mes_, getMessages().getVal(ERROR_MOVE_TITLE), JOptionPane.ERROR_MESSAGE);
         }
