+++ desktop/applis/apps/gui/gui_games/cardgamesgui/src/main/java/cards/gui/panels/PanelTricksHandsBelote.java
@@ -191,7 +191,7 @@ public class PanelTricksHandsBelote implements ViewablePanelTricksHands {
                 indice_++;
             }
             while(indice_<2*numberPlayers-1) {
-                TextLabel etiquette2_=new TextLabel(Long.toString(indice_-numberPlayers));
+                TextLabel etiquette2_=new TextLabel(Long.toString((long)indice_-numberPlayers));
                 etiquette2_.setHorizontalAlignment(SwingConstants.CENTER);
                 etiquette2_.setFont(new Font(DEFAULT,Font.BOLD,50));
                 etiquette2_.setOpaque(true);
@@ -228,7 +228,7 @@ public class PanelTricksHandsBelote implements ViewablePanelTricksHands {
                 indice_++;
             }
             while(indice_<numberPlayers*2-1) {
-                TextLabel etiquette2_=new TextLabel(Long.toString(indice_-numberPlayers));
+                TextLabel etiquette2_=new TextLabel(Long.toString((long)indice_-numberPlayers));
                 etiquette2_.setHorizontalAlignment(SwingConstants.CENTER);
                 etiquette2_.setFont(new Font(DEFAULT,Font.BOLD,50));
                 etiquette2_.setOpaque(true);
@@ -299,7 +299,7 @@ public class PanelTricksHandsBelote implements ViewablePanelTricksHands {
             }
         }
         while(indice_<2*numberPlayers-1) {
-            TextLabel etiquette2_=new TextLabel(Long.toString(indice_-numberPlayers));
+            TextLabel etiquette2_=new TextLabel(Long.toString((long)indice_-numberPlayers));
             etiquette2_.setHorizontalAlignment(SwingConstants.CENTER);
             etiquette2_.setFont(new Font(DEFAULT,Font.BOLD,50));
             etiquette2_.setOpaque(true);
